Create Main.java
This Java file has a method to generate a linked list from the .txt input file (which needs to have the proper format). After that, in the main method, we apply the reverse function we define between the m and n nodes. After that, we display the reversed linked list graphically by creating an object of the FrameDisplay class.