Create LinkedListDisplay.java
This class allows me to create the graphical view of the Linked List. In this class, we make sure that we draw a box (common way of representing a Linked List Node) with the data and the reference to the next node. The last node is the only node that has a next reference of null. We need to make sure they are in the same y coordinate but more to the right.