Create ListNode.java
This class is the definition of a Linked List Node. Every time I create a ListNode, it will have a value and a reference. Note that ListNode classes can have many more methods, but since those are not used, I tried to eliminate methods that are already known. For instance, the user enters a .txt input file that is already a linked list (which expects all nodes but the last one to have a reference to their respective next nodes).