Create fil.txt
Just like the other sample .txt files, the first number is the m node. The second node is the n node. The next row contains the linked list values. For instance, in this .txt input file I am saying, reverse the linked list between the 2nd node and the fourth node.

The row 2 3 4 6 8 9 are values that will be made into nodes that have those values. More specifically, with this input .txt file, my method creates a node with value of 2 and a next reference to the node that contains a value of 3, which in turn has a next reference to the node that contains a value of 4, which in turn contains a next reference to the node that contains the value of 6, and so on. Reversing between the 2 and 4 nodes would just reverse the 3, 4, and 6 (node that contains 3 is second node and node that contains 6 is the 4th node). Thus, graphically, my project shows the reversed linked list between the m and n nodes. It would show graphically a linked list with values              2 -> 6 -> 4 -> 3 -> 8 -> 9