Create r.txt
Just like the other sample .txt files, the first number is the m node. The second node is the n node. The next row contains the linked list values. For instance, in this .txt input file I am saying, reverse the linked list between the 3nd node and the 5th node.
3 5
1 2 3 4 5 6

The row 1 2 3 4 5 6 are values that will be made into nodes that have those values. More specifically, with this input .txt file, my method creates a node with value of 1 and a next reference to the node that contains a value of 2, which in turn has a next reference to the node that contains a value of 3, which in turn contains a next reference to the node that contains the value of 4, and so on. Reversing between the 3 and 5 nodes would just reverse the 3, 4, and 5 (node that contains 3 is third node and node that contains 5 is the fifth node). Thus, graphically, my project shows the reversed linked list between the m and n nodes. It would show graphically a linked list with values           
1 -> 2 -> 5 -> 4 -> 3 -> 6