Create s.txt
This input .txt file takes the first number which will be 2nd node. The next number is 6. This will be the 6th node. Thus, for the next row, create a linked list that has the values seen below (in the same order) and reverse that linked list between the 2nd and 6th nodes. Show it graphically. That is what my program does. More specifically, in this case, it would show 
10 -> 2 -> 19 -> 3 -> 2 -> 9 -> 1

Notice that the linked list is just reversed between the m and n nodes. The rest remains untouched.