Create l.txt
Take the first number. Take the second number. Make a linked list out of the values you see in the second row. In the exact same order. If you see a 1, then create node with value of 1. If you then see a 29. then make the node with value 1 point to a node that contains the value 29, and so on until the end. Reverse that linked list between the first number you got and the second number you got (before checking the second row). In other words, reverse the linked list between the 1st and 4th nodes.

The linked list would originally be

1 -> 29  -> 1 -> 29 -> 19 -> 0

1 would be the first

29 the second, 1 the third, and 29 the fourth

reverse them and leave the 19 and 0 untouched (they would be the 5th and 6th nodes, yet we are asked to reverse between the 1st and 4th).

Reverse and we get

29 -> 1 -> 29 -> 1 -> 19 -> 0

and show it graphically. This is what my project does.